Get tests from different sources (Selenium, fetch, and manual) and combine them. Using the parameter days, all files in automated/merged/ are updated for the last dates given the input number.

Input: Daily test data scraped via Selenium and "R fetch functions" from automated/fetch and automated/selenium directories.

Output:

automated-tests.json

countries-error.csv

which are then deployed by CI to the automated/merged/ directory in the dsbbfinddx/FINDCov19TrackerData repo.

get_test_data(days = 1, write = TRUE)

Arguments

days
days to combine. It should always be bigger than 0.

write
if csv files should be written. Default TRUE.

Details

#' When manual countries have not been updated, coronavirus_cases_new.csv has NA for tests_cumulative and new_tests. When selenium or fetch process failed and those values have not been corrected manually, coronavirus_cases_new.csv might have negative values or NA in new_tests and new_tests_corrected variables.

Countries with negative values or NA are listed as well given the days input in the folder issues/.

Parameter days allows to update past dates using manual files in manual/processed/; except for dates on "2021-02-18" or before. It allows to update from "2021-02-19" when the automatic workflow was implemented. To update data on 2021-02-18 or previous days, changes should be made using the specific date in the folder automated/merged/.

Examples

get_test_data(days = 8, write = FALSE)

#> Joining, by = c("country", "date")

#> $test_combined
#> # A tibble: 1,464 x 7
#>    country     date       tests_cumulative new_tests tests_cumulative_corrected
#>    <chr>       <date>                <dbl>     <dbl>                      <dbl>
#>  1 Afghanistan 2021-04-17               NA        NA                         NA
#>  2 Afghanistan 2021-04-18               NA        NA                         NA
#>  3 Afghanistan 2021-04-19               NA        NA                         NA
#>  4 Afghanistan 2021-04-20               NA        NA                         NA
#>  5 Afghanistan 2021-04-21               NA        NA                         NA
#>  6 Afghanistan 2021-04-22               NA        NA                         NA
#>  7 Afghanistan 2021-04-23               NA        NA                         NA
#>  8 Afghanistan 2021-04-24               NA        NA                         NA
#>  9 Albania     2021-04-17           600930      2886                     600930
#> 10 Albania     2021-04-18           603410      2480                     603410
#> # … with 1,454 more rows, and 2 more variables: new_tests_corrected <dbl>,
#> #   source <chr>
#> 
#> $countries_error
#> # A tibble: 417 x 7
#>    country     date       tests_cumulative new_tests tests_cumulative_corrected
#>    <chr>       <date>                <dbl>     <dbl>                      <dbl>
#>  1 Afghanistan 2021-04-17               NA        NA                         NA
#>  2 Afghanistan 2021-04-18               NA        NA                         NA
#>  3 Afghanistan 2021-04-19               NA        NA                         NA
#>  4 Afghanistan 2021-04-20               NA        NA                         NA
#>  5 Afghanistan 2021-04-21               NA        NA                         NA
#>  6 Afghanistan 2021-04-22               NA        NA                         NA
#>  7 Afghanistan 2021-04-23               NA        NA                         NA
#>  8 Afghanistan 2021-04-24               NA        NA                         NA
#>  9 Argentina   2021-04-17               NA        NA                         NA
#> 10 Argentina   2021-04-18               NA        NA                         NA
#> # … with 407 more rows, and 2 more variables: new_tests_corrected <dbl>,
#> #   source <chr>
#>