Get segregated tests from Selenium and combine them. Using the parameter days, all files in automated/merged/ are updated for the last dates given the input number.

Input: Daily test data scraped via Selenium automated/fetch and automated/selenium directories.

Output:

segregated.csv

which are then deployed by CI to the automated/merged/ directory in the finddx/FINDCov19TrackerData repo.

segregated_test_data(write = TRUE)

Arguments

write

if csv files should be written. Default TRUE.

Examples

segregated_test_data()
#> Rows: 161879 Columns: 8
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (3): country, jhu_ID, source
#> dbl  (4): new_tests, tests_cumulative, new_tests_corrected, tests_cumulative...
#> date (1): date
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Warning: Segregated data is scraped since 2021-03-22.
#>             Data before that date is not available
#> Joining, by = c("country", "date")
#> # A tibble: 5,032 × 15
#> # Groups:   country [10]
#>    country date       tests_cum…¹ tests…² new_t…³ source pcr_t…⁴ rapid…⁵ pcr_t…⁶
#>    <chr>   <date>           <dbl>   <dbl>   <dbl> <chr>    <dbl>   <dbl>   <dbl>
#>  1 Austria 2021-03-23    20846292  2.08e7  369807 https… 6431566  1.44e7   49167
#>  2 Austria 2021-03-24    21147134  2.11e7  300842 https… 6493316  1.47e7   61750
#>  3 Austria 2021-03-25    21565462  2.16e7  418328 https… 6568554  1.50e7   75238
#>  4 Austria 2021-03-26    22051094  2.21e7  485632 https… 6556772  1.55e7  -11782
#>  5 Austria 2021-03-27    22194436  2.22e7  143342 https… 6333726  1.59e7 -223046
#>  6 Austria 2021-03-28    22455472  2.25e7  261036 https… 6383732  1.61e7   50006
#>  7 Austria 2021-03-29    22610260  2.26e7  154788 https… 6421765  1.62e7   38033
#>  8 Austria 2021-03-30    23093306  2.31e7  483046 https… 6474269  1.66e7   52504
#>  9 Austria 2021-03-31    23451521  2.35e7  358215 https… 6543059  1.69e7   68790
#> 10 Austria 2021-04-01    23805240  2.38e7  353719 https… 6616140  1.72e7   73081
#> # … with 5,022 more rows, 6 more variables: rapid_test_new <dbl>,
#> #   percentage_rapid_daily <dbl>, pcr_test_new_corrected <dbl>,
#> #   rapid_test_new_corrected <dbl>, pcr_tests_cum_corrected <dbl>,
#> #   rapid_test_cum_corrected <dbl>, and abbreviated variable names
#> #   ¹​tests_cumulative, ²​tests_cumulative_corrected, ³​new_tests_corrected,
#> #   ⁴​pcr_tests_cum, ⁵​rapid_test_cum, ⁶​pcr_test_new