Get segregated tests from Selenium and combine them. Using the parameter days, all files in automated/merged/ are updated for the last dates given the input number.

Input: Daily test data scraped via Selenium automated/fetch and automated/selenium directories.

Output:

segregated.csv

which are then deployed by CI to the automated/merged/ directory in the dsbbfinddx/FINDCov19TrackerData repo.

segregated_test_data(write = TRUE)

Arguments

write
if csv files should be written. Default TRUE.

Examples

get_test_data(days = 8, write = FALSE)

#> Warning: Changes on 2021-02-18 or before, should made directly in automated/merged

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Joining, by = c("country", "date")

#> Rows: 97647 Columns: 7

#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (2): country, source
#> dbl  (4): tests_cumulative, new_tests, tests_cumulative_corrected, new_tests...
#> date (1): date

#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

#> $test_combined
#> # A tibble: 38,979 × 7
#>    country     date       tests_cumulative new_tests tests_cumulative_corrected
#>    <chr>       <date>                <dbl>     <dbl>                      <dbl>
#>  1 Afghanistan 2021-02-19           282249     24238                     282249
#>  2 Afghanistan 2021-02-20           282249         0                     282249
#>  3 Afghanistan 2021-02-21           282249         0                     282249
#>  4 Afghanistan 2021-02-22           282249         0                     282249
#>  5 Afghanistan 2021-02-23           282249         0                     282249
#>  6 Afghanistan 2021-02-24           282249         0                     282249
#>  7 Afghanistan 2021-02-25           282249         0                     282249
#>  8 Afghanistan 2021-02-26           282249         0                     282249
#>  9 Afghanistan 2021-02-27           282249         0                     282249
#> 10 Afghanistan 2021-02-28           282249         0                     282249
#> # … with 38,969 more rows, and 2 more variables: new_tests_corrected <dbl>,
#> #   source <chr>
#> 
#> $countries_error
#> # A tibble: 483 × 7
#> # Groups:   country [89]
#>    country     date       tests_cumulative new_tests tests_cumulative_corrected
#>    <chr>       <date>                <dbl>     <dbl>                      <dbl>
#>  1 Afghanistan 2021-09-15               NA        NA                         NA
#>  2 Afghanistan 2021-09-16               NA        NA                         NA
#>  3 Afghanistan 2021-09-17               NA        NA                         NA
#>  4 Afghanistan 2021-09-18               NA        NA                         NA
#>  5 Afghanistan 2021-09-19               NA        NA                         NA
#>  6 Andorra     2021-07-30           208859   -158093                     208859
#>  7 Argentina   2021-09-15               NA        NA                         NA
#>  8 Argentina   2021-09-16               NA        NA                         NA
#>  9 Argentina   2021-09-17               NA        NA                         NA
#> 10 Argentina   2021-09-18               NA        NA                         NA
#> # … with 473 more rows, and 2 more variables: new_tests_corrected <dbl>,
#> #   source <chr>
#>